AIinterntask/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pdfplumber
python-docx
pytesseract
opencv-python  # Removed version pin to resolve conflict
Pillow
paddleocr
paddlepaddle==3.0.0
pymupdf
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.31.0
chromadb==0.4.24
sentence-transformers==2.7.0
psutil==5.9.8
streamlit